search input is well aligned
